delete interfaces subfolder and move files into individual submodules, rename methods submodule to bias_estimator, rename bias submodule to bias_creator
