+++ src/inhomcorr/bias_creator/bias_creator_interface.py
@@ -2,7 +2,7 @@
 from abc import ABC
 from abc import abstractmethod
 
-from inhomcorr.interfaces import ImageData
+from inhomcorr.mrdata import ImageData
 
 
 class BiasCreator(ABC):
