+++ src/inhomcorr/bias_estimator/supervised_training_dataset_interface.py
@@ -4,7 +4,7 @@ from abc import ABC
 from torch.utils.data import Dataset
 
 
-class BiasFieldDataset(ABC, Dataset):
+class SupervisedTrainingDataset(ABC, Dataset):
     """Bias Field Dataset Interface."""
 
     pass
