+++ src/inhomcorr/data_loader/data_loader_interface.py
@@ -2,7 +2,7 @@
 from abc import ABC
 from abc import abstractmethod
 
-from inhomcorr.interfaces import QMRIData
+from inhomcorr.mrdata import QMRIData
 
 
 class QMRIDataLoader(ABC):
