--- src/inhomcorr/interfaces/__init__.py
@@ -1,10 +1,0 @@
-from .mr_data_interface import ImageData as ImageData  # isort:skip
-from .mr_data_interface import QMRIData as QMRIData  # isort:skip
-from .bias_creator_interface import BiasCreator
-from .bias_estimator_interface import BiasEstimator
-from .bias_estimator_interface import HyperParameters
-from .dataset_interface import BiasFieldDataset
-from .mr_sig_interface import MRParam as MRParam
-from .mr_sig_interface import MRSig as MRSig
-from .trainer_interface import Trainer
-from .trainer_interface import TrainingParameters
