+++ tests/test_mr_sig.py
@@ -1,7 +1,7 @@
 """Tests of mr signal T1 object."""
 import unittest
 
-from inhomcorr.interfaces.mr_data_interface import ImageData
+from inhomcorr.mrdata import ImageData
 from inhomcorr.mrsig.flash import MRSigFlash
 from tests.testdata import TestData
